Kernel/Net: Move Intel network adapters code to a separate directory
