+++ Kernel/Net/Intel/E1000ENetworkAdapter.h
@@ -12,7 +12,7 @@
 #include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/Bus/PCI/Device.h>
 #include <Kernel/Interrupts/IRQHandler.h>
-#include <Kernel/Net/E1000NetworkAdapter.h>
+#include <Kernel/Net/Intel/E1000NetworkAdapter.h>
 #include <Kernel/Net/NetworkAdapter.h>
 #include <Kernel/Random.h>
