+++ Kernel/Net/Intel/E1000NetworkAdapter.h
